De-duplicate lintWithFeatherRule and HTTP socket utilities
