+++ src/cli/test/status-server-stop.test.ts
@@ -1,9 +1,9 @@
 import assert from "node:assert/strict";
-import net from "node:net";
 import { describe, it } from "node:test";
 import { setTimeout as delay } from "node:timers/promises";
 
 import { startStatusServer } from "../src/modules/status/server.js";
+import { createHttpSocketAndWaitForResponse } from "./test-helpers/http-socket-utils.js";
 
 void describe("status server lifecycle", () => {
     void it("closes active sockets on stop", async () => {
@@ -20,31 +20,11 @@ void describe("status server lifecycle", () => {
             })
         });
 
-        const socket = net.createConnection({
-            host: controller.host,
-            port: controller.port
-        });
-        socket.setEncoding("utf8");
-
-        const closePromise = new Promise<void>((resolve) => {
-            socket.once("close", () => resolve());
-        });
-
-        const responsePromise = new Promise<void>((resolve, reject) => {
-            let buffer = "";
-            const handleData = (chunk: string) => {
-                buffer += chunk;
-                if (buffer.includes("\r\n\r\n")) {
-                    socket.off("data", handleData);
-                    resolve();
-                }
-            };
-
-            socket.on("data", handleData);
-            socket.once("error", reject);
-        });
-
-        socket.write("GET /status HTTP/1.1\r\nHost: localhost\r\nConnection: keep-alive\r\n\r\n");
+        const { socket, closePromise, responsePromise } = await createHttpSocketAndWaitForResponse(
+            controller.host,
+            controller.port,
+            "GET /status HTTP/1.1\r\nHost: localhost\r\nConnection: keep-alive\r\n\r\n"
+        );
 
         await responsePromise;
         await controller.stop();
