+++ src/lint/test/feather-plugin-fixture-migration.test.ts
@@ -7,7 +7,7 @@ import { fileURLToPath } from "node:url";
 
 import * as LintWorkspace from "@gml-modules/lint";
 
-import { applyFixOperations, createLocResolver, type ReplaceTextRangeFixOperation } from "./rule-test-harness.js";
+import { lintWithFeatherRule as runFeatherRule } from "./rule-test-harness.js";
 
 type MigrationCase = {
     fixtureDirectory: string;
@@ -33,46 +33,7 @@ function lintWithFeatherRule(
     ruleName: string,
     code: string
 ): { messages: Array<{ messageId: string }>; output: string } {
-    const rule = LintWorkspace.Lint.plugin.rules[ruleName];
-    const messages: Array<{ messageId: string; fix?: ReplaceTextRangeFixOperation }> = [];
-    const getLocFromIndex = createLocResolver(code);
-
-    const context = {
-        options: [{}],
-        sourceCode: {
-            text: code,
-            getLocFromIndex
-        },
-        report(payload: {
-            messageId: string;
-            fix?: (fixer: {
-                replaceTextRange(range: [number, number], text: string): ReplaceTextRangeFixOperation;
-            }) => ReplaceTextRangeFixOperation | null;
-        }) {
-            const fixer = {
-                replaceTextRange(range: [number, number], text: string): ReplaceTextRangeFixOperation {
-                    return { kind: "replace", range, text };
-                }
-            };
-            messages.push({
-                messageId: payload.messageId,
-                fix: payload.fix ? (payload.fix(fixer) ?? undefined) : undefined
-            });
-        }
-    } as never;
-
-    const listeners = rule.create(context);
-    listeners.Program?.({ type: "Program" } as never);
-
-    const output = applyFixOperations(
-        code,
-        messages.map((message) => message.fix).filter((fix): fix is ReplaceTextRangeFixOperation => fix !== undefined)
-    );
-
-    return {
-        messages: messages.map((message) => ({ messageId: message.messageId })),
-        output
-    };
+    return runFeatherRule(LintWorkspace.Lint.plugin, ruleName, code);
 }
 
 async function readMigratedFeatherFixture(fixtureDirectory: string): Promise<string> {
