+++ src/lint/test/rule-test-harness.ts
@@ -68,3 +68,54 @@ export function applyFixOperations(text: string, operations: Array<RuleTestFixOp
     output += text.slice(cursor);
     return output;
 }
+
+/**
+ * Runs a feather rule against code and returns the messages and fixed output.
+ * This is a shared test helper for testing feather lint rules.
+ */
+export function lintWithFeatherRule(
+    plugin: { rules: Record<string, { create: (context: never) => { Program?: (node: never) => void } }> },
+    ruleName: string,
+    code: string
+): { messages: Array<{ messageId: string }>; output: string } {
+    const rule = plugin.rules[ruleName];
+    const messages: Array<{ messageId: string; fix?: ReplaceTextRangeFixOperation }> = [];
+    const getLocFromIndex = createLocResolver(code);
+
+    const context = {
+        options: [{}],
+        sourceCode: {
+            text: code,
+            getLocFromIndex
+        },
+        report(payload: {
+            messageId: string;
+            fix?: (fixer: {
+                replaceTextRange(range: [number, number], text: string): ReplaceTextRangeFixOperation;
+            }) => ReplaceTextRangeFixOperation | null;
+        }) {
+            const fixer = {
+                replaceTextRange(range: [number, number], text: string): ReplaceTextRangeFixOperation {
+                    return { kind: "replace", range, text };
+                }
+            };
+            messages.push({
+                messageId: payload.messageId,
+                fix: payload.fix ? (payload.fix(fixer) ?? undefined) : undefined
+            });
+        }
+    } as never;
+
+    const listeners = rule.create(context);
+    listeners.Program?.({ type: "Program" } as never);
+
+    const output = applyFixOperations(
+        code,
+        messages.map((message) => message.fix).filter((fix): fix is ReplaceTextRangeFixOperation => fix !== undefined)
+    );
+
+    return {
+        messages: messages.map((message) => ({ messageId: message.messageId })),
+        output
+    };
+}
